+++ src/commands/Dashboard.ts
@@ -9,7 +9,7 @@ import { DashboardData } from '../models/DashboardData';
 import { MediaLibrary } from '../helpers/MediaLibrary';
 import { DashboardListener, MediaListener, SettingsListener, TelemetryListener, DataListener, PagesListener, ExtensionListener, SnippetListener, TaxonomyListener } from '../listeners/dashboard';
 import { MediaListener as PanelMediaListener } from '../listeners/panel'
-import { ModeListener } from '../listeners/general';
+import { GitListener, ModeListener } from '../listeners/general';
 
 export class Dashboard {
   private static webview: WebviewPanel | null = null;
@@ -146,6 +146,7 @@ export class Dashboard {
       TelemetryListener.process(msg);
       SnippetListener.process(msg);
       ModeListener.process(msg);
+      GitListener.process(msg);
       TaxonomyListener.process(msg);
     });
   }