+++ src/constants/settings.ts
@@ -78,6 +78,9 @@ export const SETTING_FRAMEWORK_START = "framework.startCommand";
 
 export const SETTING_SITE_BASEURL = "site.baseURL";
 
+export const SETTING_GIT_ACTIONS = "git.actions";
+export const SETTING_GIT_COMMIT_MSG = "git.commitMsg";
+
 /**
  * @deprecated
  */